+++ R/npars.R
@@ -7,7 +7,6 @@ universals::npars
 #'
 #' @examples
 #' npars(as.term(c("sigma", "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]")))
-#' npars(as.term(c("sigma", "alpha[1]", "alpha[2]")), scalar = TRUE)
 npars.term <- function(x, scalar = NA, ...) {
   chk_unused(...)
   if (anyNA(x)) {